+++ example/lib/ui/style.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+const textStyle = TextStyle(
+  fontSize: 12,
+  color: Colors.black,
+  decoration: TextDecoration.none,
+  fontWeight: FontWeight.normal,
+  fontFamily: 'Roboto'
+);