+++ lib/src/debug/debug_w.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import '../calculations/calculator.dart';
 import '../models/tile_point.dart';
 import '../ui/helper.dart';
+import '../ui/style.dart';
 
 /// DebugW
 class DebugW extends StatelessWidget {
@@ -30,33 +31,40 @@ class DebugW extends StatelessWidget {
     final centerTile =
         getCenterTile(center: center, zoom: zoom, mapScale: mapScale);
 
-    final lZoom = (zoom * 1000000).toInt() / 1000000;
-    final lMapScale = (mapScale * 1000).toInt() / 1000;
+    Widget item(String title, String value, bool isLeft) {
+      final align = isLeft ? Alignment.centerLeft : Alignment.centerRight;
+      Widget curr = Text('$title: $value', style: textStyle);
+      curr = FittedBox(fit: BoxFit.scaleDown, child: curr);
+      curr = Align(alignment: align, child: curr);
+      return Expanded(child: curr);
+    }
 
     Widget current = Column(
       children: [
         Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Text('x: ${center.x}'),
-            Text('Tiles: ${loadedTiles.length}'),
+            item('x', center.x.format(), true),
+            item('Tiles', loadedTiles.length.toString(), false),
           ],
         ),
         Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Text('y: ${center.y}'),
-            Text('Zoom: $lZoom'),
+            item('y', center.y.format(), true),
+            item('Zoom', zoom.format(), false),
           ],
         ),
         Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Text('Center Tile: $centerTile'),
-            Text('MapScale: $lMapScale'),
+            item('CT', centerTile.toString(), true),
+            item('Scale', mapScale.format(), false),
           ],
         ),
       ],
+    );
+
+    current = Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 4),
+      child: current,
     );
 
     current = SizedBox(width: fullW, child: current);
@@ -64,5 +72,15 @@ class DebugW extends StatelessWidget {
     current = Positioned(top: 0, left: 0, child: current);
 
     return current;
+  }
+}
+
+extension SomeExtension on double {
+  String format() {
+    var val = toStringAsFixed(3);
+    while (val.endsWith('0') || val.endsWith('.')) {
+      val = val.replaceRange(val.length - 1, val.length, '');
+    }
+    return val;
   }
 }